music and sfx

+++ IT008_Game/Game/GameObjects/IEnemy.cs
@@ -6,6 +6,6 @@ namespace IT008_Game.Game.GameObjects
     {
         public int GetWeight();
         public Sprite2D GetSprite();
-        public void Damage(int damage);
+        public void Damage(float damage);
     }
 }
